(++fix prev commit)repo detail page
Create another component to represent the repo detail page. Once the UI is created for this page, add a route to dynamically pass the data of the repo to be displayed. Also add a convenience method to the collection to find the appropriate repo. (reverted from commit ed8d811d2f57743afa3fc514f17ae419a9c78015)